Optimized filters for equality
As we scan through the virtual memory, unless we have chosen a ridiculous filter, most of the addresses are going to be a miss.

Suppose we are comparing whether an address contains a value that is equal to an integer value.
If the address is not equal to the value as an Int8, then it makes no sense to test for Int16, Int32 and Int64.

Before, Filters were of type [Type] -> ByteString -> Bool,

Now, Filters are of type ByteString -> [Type].

This way, we can create a filter

eqInt:: ByteString -> [Type]

that does the following:

Compare the address to a value as an Word8,
if compatible, then compare it as an Word16... and so on.

But since most of the time, the address is NOT going to be equal to the word8, most of the scan is going to consist of a simple Word8 comparison in eqBSN.

So scanning for [Int64, Int32, Int16, Int8] is going to take roughly the same time as scanning for Int8.