FilterInfo is now OP. Filter is lame
Since we change Filter to be
Filter :: ByteString -> [Data]

we had no way of calculating max_size_of for loading of chunks in PeepingTom.State.

To get this information, we now provide a FilterInfo instead of a Filter to the various functions in PeepingTom.State.

FilterInfo is simply defined as

type FilterInfo = (Filter, Int)

so we are just as fast, and we can properly allocate space for chunks.