Added some new filters, Fixed error message
Some new filters both to filter regions as well to filter candidates.

The new RFilters do not work. Sometimes, we are uncapable of reading from a section of memory (even though it has Read permissions set).

Currently, Posix.pread just throws an exception. We should probably handle this somewhere else actually.
Probably just ignore it, and continue to the next region when this happens.