Added the possibility to search/write strings
Now we can search for strings.

This comes in the form of Commands.StrSet in the application.

In the library, this comes in the form of:

1. Fixing strFilter.
2. A new strWriter, as well as bytesWriter and byterWriter_

Posix no longer throws an exception. Not memory chunks have a mcOk, which is true when the memory read was ok.
When we scan regions, now it checks whether the read was ok before, and if it wasn't, it simply outputs the error and continues on.