Progress
Some progress.
We now have a working eqInt, as well as i8Eq, i16Eq, i32Eq and i64Eq.
It seems to match the results of ScanMem and regular PeepingTom.

And the performances is still about 3.3x the speed of ScanMem, which is a significant improval from the 17x from before.

Now, I need to move things around, implement the ByteString filters,
and then maybe some naming changes across the project